sriov: Add a test to check iommu support
This PR adds '[Capabilities] check virsh capabilities output
includes iommu support element' case.
