+++ libvirt/tests/cfg/sriov/capabilities/sriov_capabilities_iommu_support.cfg
@@ -1,0 +1,4 @@
+- sriov.capabilities.iommu_support:
+    type = sriov_capabilities_iommu_support
+    start_vm = "no"
+    only x86_64
