--- libvirt/tests/cfg/sriov/sriov_capabilities.cfg
@@ -1,6 +1,0 @@
-- sriov.capabilities:
-    type = sriov_capabilities
-    start_vm = "no"
-    only x86_64
-    variants test_case:
-        - intel_iommu:
